+++ Centaurus.Test.Domain/BaseQuantumHandlerTests.cs
@@ -184,7 +184,7 @@ namespace Centaurus.Test
 
             var account = Global.AccountStorage.GetAccount(clientKeyPair);
             if (requestLimit.HasValue)
-                account.Account.RequestRateLimits = new RequestRateLimits { HourLimit = requestLimit.Value, MinuteLimit = requestLimit.Value };
+                account.Account.RequestRateLimits = new RequestRateLimits { HourLimit = (uint)requestLimit.Value, MinuteLimit = (uint)requestLimit.Value };
 
             var minuteLimit = (account.Account.RequestRateLimits ?? Global.Constellation.RequestRateLimits).MinuteLimit;
             var minuteIterCount = minuteLimit + 1;
